% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{insertPlotlyReset}
\alias{insertPlotlyReset}
\title{Adds a js$reset_<id>_<event>() to reset the selection on a plot}
\usage{
insertPlotlyReset(source, event = c("hover", "click", "selected"))
}
\description{
selected elements stick on a plot even after it is redrawn, most of the
times you don't want to do that.
}
\details{
See:
https://community.plot.ly/t/reseting-click-events/2718/2
https://stackoverflow.com/questions/44412382/
}
