% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-reactiveGeneSetDb-module.R
\name{reactiveGeneSetDbFilterUI}
\alias{reactiveGeneSetDbFilterUI}
\title{UI components to filter a GeneSetDb by collection and set size}
\usage{
reactiveGeneSetDbFilterUI(id, min = 2, max = 100L, ...)
}
\description{
UI components to filter a GeneSetDb by collection and set size
}
