% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-reactiveGeneSetSelect.R
\name{reactiveGeneSetSelect}
\alias{reactiveGeneSetSelect}
\title{A geneset selector for a ReactiveGeneSetDb}
\usage{
reactiveGeneSetSelect(input, output, session, gdb = NULL, ...)
}
\arguments{
\item{gdb}{A [reactiveGeneSetDb()] module, or `NULL`}
}
\value{
A list of geneset info and membersihp
}
\description{
This module is unique becauase if you pass `gdb = NULL` it returns a shim
of iteself that will never trigger reactivity.
}
