% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-userDefinedGeneSetDb.R
\name{userDefinedGeneSetDb}
\alias{userDefinedGeneSetDb}
\title{Creates a GeneSetDb from a user-specfied gene set definition table.}
\usage{
userDefinedGeneSetDb(input, output, session, ...)
}
\value{
A list of reactive components. `$gdb()` will be a GeneSetDb when
  the user uploades a valid gene set definition file. Otherwise it will be
  `NULL`.
}
\description{
This module provides an upload button that allows a user to upload a
table of gene set definitions in [mutiGSEA::GeneSetDb()] format. Minimal
validation checks are implemented.
}
